% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{install_use_push}
\alias{install_use_push}
\title{Convenience function to run full testing}
\usage{
install_use_push(ref = "main")
}
\description{
Convenience function to run full testing
}
\keyword{internal}
