% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{wait_long_for_head}
\alias{wait_long_for_head}
\title{This function brings a big speed improvement for the PR}
\usage{
wait_long_for_head(wait)
}
\description{
This function brings a big speed improvement for the PR
}
